+++ inst/tinytest/test-hypothesis.R
@@ -437,4 +437,12 @@ expect_equivalent(h1$p.value, h2$test$pvalues)
 expect_equivalent(coef(mod)[2:3] + 3.5, h1$estimate)
 
 
+# Issue #1453: ratio ~ should use hypothesis=1 as null
+mod <- lm(mpg ~ factor(gear) - 1, data = mtcars)
+h1 <- hypotheses(mod, hypothesis = ratio ~ sequential)
+h2 <- hypotheses(mod, hypothesis = c("b2 / b1 = 1", "b3 / b2 = 1"))
+expect_equivalent(h1$p.value, c(0.000243703399238325, 0.191786862518089))
+expect_equivalent(h2$p.value, c(0.000243703399238325, 0.191786862518089))
+
+
 rm(list = ls())